+++ src/prefect/_internal/schemas/bases.py
@@ -10,7 +10,7 @@ from pydantic import BaseModel, ConfigDict, Field
 from rich.repr import RichReprResult
 from typing_extensions import Self
 
-from prefect.types import uuid7
+from prefect._internal.uuid7 import uuid7
 from prefect.types._datetime import (
     DateTime,
     human_friendly_diff,
