+++ src/prefect/task_runners.py
@@ -21,6 +21,7 @@ from typing import (
 
 from typing_extensions import ParamSpec, Self, TypeVar
 
+from prefect._internal.uuid7 import uuid7
 from prefect.client.schemas.objects import TaskRunInput
 from prefect.exceptions import MappingLengthMismatch, MappingMissingIterable
 from prefect.futures import (
@@ -31,7 +32,6 @@ from prefect.futures import (
 )
 from prefect.logging.loggers import get_logger, get_run_logger
 from prefect.settings import PREFECT_TASK_RUNNER_THREAD_POOL_MAX_WORKERS
-from prefect.types import uuid7
 from prefect.utilities.annotations import allow_failure, quote, unmapped
 from prefect.utilities.callables import (
     collapse_variadic_parameters,
