+++ src/prefect/types/__init__.py
@@ -6,7 +6,6 @@ from uuid import UUID
 from typing_extensions import Literal
 import orjson
 import pydantic
-from uuid_extensions import uuid7 as _uuid7  # pyright: ignore[reportMissingTypeStubs]
 
 from ._datetime import DateTime, Date
 from .names import (
@@ -154,10 +153,6 @@ KeyValueLabelsField = Annotated[
 ]
 
 
-def uuid7() -> UUID:
-    return cast(UUID, _uuid7())
-
-
 __all__ = [
     "BANNED_CHARACTERS",
     "WITHOUT_BANNED_CHARACTERS",
@@ -178,5 +173,4 @@ __all__ = [
     "SecretDict",
     "StatusCode",
     "StrictVariableValue",
-    "uuid7",
 ]
